+++ src/PortalBase/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add log message code `1637421327` to `\Heptacom\HeptaConnect\Portal\Base\Exploration\ExplorerStack::getOrigin` when a short-notation explorer is unconfigured
+- Add log message code `1637607487` to `\Heptacom\HeptaConnect\Portal\Base\Reception\ReceiverStack::getOrigin` when a short-notation explorer is unconfigured
+- Add log message code `1637607653` to `\Heptacom\HeptaConnect\Portal\Base\Emission\EmitterStack::getOrigin` when a short-notation explorer is unconfigured
+- Add log message code `1637607699` to `\Heptacom\HeptaConnect\Portal\Base\Web\Http\HttpHandlerStack::getOrigin` when a short-notation explorer is unconfigured
+
 ### Changed
 
 - Use container tags `heptaconnect.flow_component.status_reporter_source`, `heptaconnect.flow_component.emitter_source`, `heptaconnect.flow_component.explorer_source`, `heptaconnect.flow_component.receiver_source`, `heptaconnect.flow_component.web_http_handler_source` instead of `heptaconnect.flow_component.emitter`, `heptaconnect.flow_component.emitter_decorator`, `heptaconnect.flow_component.explorer`, `heptaconnect.flow_component.explorer_decorator`, `heptaconnect.flow_component.receiver`, `heptaconnect.flow_component.receiver_decorator` and `heptaconnect.flow_component.web_http_handler` to collect flow component services